created io.h, removed inb and outb from serial.h
