+++ kernel/serial.h
@@ -4,17 +4,9 @@
 #include <stddef.h>
 #include <stdint.h>
 
+#include "io.h"
+
 #define COM1 0x3F8
-
-static inline void outb(uint16_t port, uint8_t val) {
-    asm volatile ("outb %0, %1" : : "a"(val), "Nd"(port));
-}
-
-static inline uint8_t inb(uint16_t port) {
-    uint8_t ret;
-    asm volatile ("inb %1, %0" : "=a"(ret) : "Nd"(port));
-    return ret;
-}
 
 static inline void serial_init() {
     outb(COM1 + 1, 0x00); // Disable interrupts
